+++ db/migrate/20180808190054_create_products_sales.rb
@@ -3,12 +3,12 @@ class CreateProductsSales < ActiveRecord::Migration[5.2]
     create_table :products_sales do |t|
       t.integer :sales_transcation_id
       t.integer :product_id
-      t.integer :qty
-      t.integer :total
-      t.integer :total_saving
+      t.decimal :qty
+      t.decimal :total
+      t.decimal :total_saving
       t.string  :product_name
-      t.integer :retail_price
-      t.integer :pomo_price
+      t.decimal :retail_price
+      t.decimal :pomo_price
 
 
       t.datetime :created_at, null: false
